requirements.txt
streamlit
groq
razorpay
openai
firebase-admin
pypdf
fpdf
gTTS
google-generativeai